+++ reqcoco-api/src/main/java/net/paissad/tools/reqcoco/api/model/Status.java
@@ -1,7 +1,15 @@
 package net.paissad.tools.reqcoco.api.model;
 
+import lombok.Getter;
+
 public enum Status {
 
-    TODO, IN_PROGRESS, DONE, FAILED, IGNORE;
+    TODO("TODO"), IN_PROGRESS("IN_PROGRESS"), DONE("DONE"), FAILED("FAILED"), IGNORE("IGNORED");
 
+    @Getter
+    private final String displayName;
+
+    private Status(final String displayName) {
+        this.displayName = displayName;
+    }
 }